zcash_pool_migration: Split the satisfiability machinery into its own module
`engine` had grown to ~7.6k lines by carrying three distinct concerns:
planning and committing a migration, the vocabulary a store answers its
liveness questions in, and the drive loop built on those answers. Move the
latter two to `zcash_pool_migration::satisfiability`.

The oracle vocabulary (`StepSatisfiability`, `UnsatisfiableCause`,
`UnsatisfiableKind`, `ParseUnsatisfiableKindError`, `InputObservation`,
`classify_input_observations`) and the caller policies those answers are judged
under (`ReorgSettleDepth`, `ReplanThreshold`) move together with the verified
drive API (`advance_migration`, `AdvanceConfig`, `DuenessTargets`, and the
private report-adjudication and broadening helpers) and the drive API's tests.
`engine` keeps planning, building, signing, proving, and rebuilding, and the
persisted types those produce; `state` keeps the planning kernel.

No `pub use` re-exports: each moved item has exactly one path. Every consumer
imports from the new module instead, and the `Unreleased` CHANGELOG entries name
the paths this crate will first publish.

Also record why `MigrationTransaction::spend_nullifiers` holds raw `[u8; 32]`
rather than `orchard::note::Nullifier`, cross-referencing the same argument made
at `lock_owner`.
