+++ zcash_pool_migration/src/satisfiability.rs
@@ -1,0 +1,1958 @@
+//! Whether a committed migration's pre-signed transactions can still execute, and the drive loop
+//! that acts on the answer.
+//!
+//! Two things live here. The first is the VOCABULARY a store answers in: [`StepSatisfiability`]
+//! and its [`UnsatisfiableCause`]s, the reduced [`UnsatisfiableKind`] persisted beside a mark, the
+//! per-nullifier [`InputObservation`]s a store folds through [`classify_input_observations`], and
+//! the caller policies ([`ReorgSettleDepth`], [`ReplanThreshold`]) those answers are judged under.
+//! A store implements
+//! [`check_step_satisfiability`](crate::engine::PoolMigrationRead::check_step_satisfiability) by
+//! supplying data access alone; what an observation MEANS is decided once, here. The second is
+//! [`advance_migration`], the verified drive API a consuming application turns the crank with: it
+//! asks the [`state`](crate::state) kernel what to do next from the persisted state alone, puts the
+//! transaction the kernel names to the store's oracle, records and propagates whatever
+//! determinations that turns up, and surfaces only a step the wallet has vouched for. So
+//! [`engine`](crate::engine) plans, builds, and commits a migration, `state` decides what the
+//! committed form permits next, and this module is where those decisions meet the wallet's live
+//! view of the chain.
+
+use alloc::vec::Vec;
+use core::fmt;
+
+use zcash_protocol::consensus::BlockHeight;
+
+use crate::engine::{
+    MigrationState, MigrationTransferId, MigrationTxState, PoolMigrationRead, PoolMigrationWrite,
+};
+use crate::state::AdvanceStep;
+
+/// The share of planned transfer value (an integer percent) above which a migration with
+/// unsatisfiable transfers should be re-planned IMMEDIATELY rather than after satisfiable work
+/// drains. Stamped on the migration at commit (like the anchor bucket grid), so every consumer
+/// of the same state applies the same policy: the threshold governs a determination derived
+/// from persisted marks, and two consumers of one migration must not disagree about it.
+///
+/// The two endpoints have distinct meanings: `0` triggers a replan as soon as ANY value is marked
+/// unsatisfiable, however small; `100` never triggers immediately (the strict `>` in
+/// [`MigrationState::replan_required`] can never be satisfied), so an unsatisfiable migration is
+/// only ever re-planned once satisfiable work drains.
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub struct ReplanThreshold(u8);
+
+impl ReplanThreshold {
+    /// The default policy: strictly more than 20% of planned transfer value unsatisfiable
+    /// triggers an immediate replan.
+    pub const DEFAULT: Self = Self(20);
+
+    /// A threshold of `percent` (0 ..= 100), or `None` if `percent > 100`.
+    pub const fn new(percent: u8) -> Option<Self> {
+        if percent <= 100 {
+            Some(Self(percent))
+        } else {
+            None
+        }
+    }
+
+    /// The integer percent.
+    pub const fn percent(&self) -> u8 {
+        self.0
+    }
+}
+
+/// How many blocks the chain must advance past a divergence before a displacement is treated as
+/// PERMANENT. Caller policy, supplied to [`PoolMigrationRead::check_step_satisfiability`]: the
+/// right value tracks the chain's block spacing (on the order of the sync wake-up settle margin —
+/// minutes, not hours — at the current 75-second target), and keeping it current is the caller's
+/// responsibility; this crate deliberately ships no default and never persists it. A displacement
+/// judged settled under an aggressive depth self-corrects: the marks it produces are cleared by
+/// reorg truncation if the chain swings back.
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub struct ReorgSettleDepth(u32);
+
+impl ReorgSettleDepth {
+    /// A settle depth of `blocks` blocks.
+    pub const fn new(blocks: u32) -> Self {
+        Self(blocks)
+    }
+
+    /// The depth in blocks.
+    pub const fn blocks(&self) -> u32 {
+        self.0
+    }
+}
+
+/// A store's answer to "does the environment this store lives in obstruct this pre-signed
+/// transaction?" — see [`PoolMigrationRead::check_step_satisfiability`]. Deliberately EXHAUSTIVE:
+/// executable now, not yet, or never is a complete classification of a step's disposition, and
+/// the consumers' cause-dependent responses are compiler-forced over it.
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub enum StepSatisfiability {
+    /// Every queried input is among the account's unspent notes, and nothing else obstructs.
+    Satisfiable {
+        /// The fully-scanned height the observation rests on.
+        as_of_height: BlockHeight,
+    },
+    /// Some input's nullifier corresponds to no note the wallet has seen — its source is not yet
+    /// mined, or not yet scanned — or a known creator is unmined but still viable. Not an
+    /// obstruction: retry after further sync.
+    NotYetSatisfiable {
+        /// The fully-scanned height the observation rests on.
+        as_of_height: BlockHeight,
+    },
+    /// Permanently obstructed.
+    Unsatisfiable {
+        /// Why the step can never execute.
+        cause: UnsatisfiableCause,
+        /// The fully-scanned height the observation rests on.
+        as_of_height: BlockHeight,
+    },
+}
+
+/// Why a step can never execute. `#[non_exhaustive]`: new reasons a step can never execute (a
+/// retention-grid change, note-lock conflicts) have a home without breaking consumers' matches.
+#[non_exhaustive]
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub enum UnsatisfiableCause {
+    /// The wallet has seen these inputs spent: their nullifiers were recognized in mined
+    /// transactions.
+    InputsSpent {
+        /// The affected inputs' nullifiers, in the cache's action order.
+        nullifiers: Vec<[u8; 32]>,
+    },
+    /// The wallet knows notes for these inputs, but their creating transaction is unmined and can
+    /// never mine: it was proven against an anchor a settled reorg permanently invalidated. The
+    /// inputs will never exist on chain. Carries the invalidated anchor (raw root bytes) —
+    /// evidence the caller cannot recover from the checked transaction itself, since it belongs
+    /// to the producer.
+    InputsInvalidated {
+        /// The dead anchor's raw root bytes.
+        anchor: [u8; 32],
+    },
+    /// The transaction's expiry height has passed without it mining (ZIP 203): the pre-signed
+    /// artifact can never be included in a block.
+    Expired,
+    /// The transaction is broadcast but unmined, its installed anchor is no longer a root on the
+    /// current chain, and the invalidating reorg is settled per the caller's
+    /// [`ReorgSettleDepth`]: as proven, it can never be mined.
+    AnchorInvalidated,
+}
+
+impl UnsatisfiableCause {
+    /// The [`UnsatisfiableKind`] recorded alongside the mark this cause applies, or `None` for a
+    /// cause that applies no mark ([`Expired`](Self::Expired), which only confirms a derivation
+    /// the kernel already makes from the same
+    /// [`expiry_height`](crate::engine::MigrationTransaction::expiry_height)).
+    ///
+    /// This is the single definition of BOTH decisions — whether to mark, and under which kind —
+    /// so the stamp and the kind stored beside it cannot disagree about whether there is a mark at
+    /// all. [`Inherited`](UnsatisfiableKind::Inherited) is never returned here: it names a mark
+    /// that arrived through the dependency closure rather than from any observed cause.
+    pub const fn kind(&self) -> Option<UnsatisfiableKind> {
+        // Compiler-forced over the (in-crate exhaustive) cause vocabulary: a new cause must decide
+        // here whether it is an observation to store — and under which kind — or a derivation to
+        // confirm.
+        match self {
+            UnsatisfiableCause::InputsSpent { .. } => Some(UnsatisfiableKind::InputsSpent),
+            UnsatisfiableCause::InputsInvalidated { .. } => {
+                Some(UnsatisfiableKind::InputsInvalidated)
+            }
+            UnsatisfiableCause::AnchorInvalidated => Some(UnsatisfiableKind::AnchorInvalidated),
+            UnsatisfiableCause::Expired => None,
+        }
+    }
+
+    /// Whether this cause records a DURABLE mark: it reports chain state the state machine cannot
+    /// re-derive (the input-level causes, and the anchor-level one), as opposed to
+    /// [`Expired`](Self::Expired), which only confirms a derivation the kernel already makes from
+    /// the same [`expiry_height`](crate::engine::MigrationTransaction::expiry_height).
+    ///
+    /// Derived from [`kind`](Self::kind) rather than restating its match, deliberately: both
+    /// [`MigrationState::record_satisfiability`] (which applies the mark) and
+    /// [`advance_migration`] (which decides whether an answer is a discovery worth recording and
+    /// broadening on) must agree, because the drive loop's TERMINATION rests on their agreement —
+    /// each recorded discovery must mark at least the candidate that produced it, or the loop
+    /// could re-plan onto the same candidate forever. A second copy of this match could drift into
+    /// exactly that hang.
+    pub(crate) const fn marks(&self) -> bool {
+        self.kind().is_some()
+    }
+}
+
+/// WHY a transaction can never mine, as persisted beside its
+/// [`unsatisfiable_at`](crate::engine::MigrationTransaction::unsatisfiable_at) stamp and rendered
+/// by a wallet: a reduced discriminant of the [`UnsatisfiableCause`] that produced the mark, plus
+/// [`Inherited`](Self::Inherited) for a mark the dependency closure applied.
+///
+/// The full cause is deliberately NOT persisted. Its variants carry evidence payloads — a spent
+/// nullifier list, a dead anchor root — that nothing in the state machine ever reads back, and
+/// that a store can re-derive by asking its oracle again; storing them would grow every marked row
+/// by an unbounded list to answer a question consumers ask only in words. What a consumer needs
+/// durably is which of these happened, which is what this records.
+///
+/// `#[non_exhaustive]`: this tracks the cause vocabulary, which is itself open, so a new marking
+/// cause is expected to bring a new kind with it.
+#[non_exhaustive]
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub enum UnsatisfiableKind {
+    /// The wallet saw this transaction's own inputs spent: their nullifiers were recognized in
+    /// mined transactions ([`UnsatisfiableCause::InputsSpent`]).
+    InputsSpent,
+    /// This transaction's own inputs will never exist on chain: the wallet knows notes for them,
+    /// but their creating transaction is unmined and was proven against an anchor a settled reorg
+    /// permanently invalidated ([`UnsatisfiableCause::InputsInvalidated`]).
+    InputsInvalidated,
+    /// This transaction is broadcast and unmined, and the anchor its OWN proof installed is no
+    /// longer a root of the chain the wallet has scanned
+    /// ([`UnsatisfiableCause::AnchorInvalidated`]).
+    AnchorInvalidated,
+    /// Nothing was observed about this transaction itself: the mark arrived through the DEPENDENCY
+    /// CLOSURE (see [`MigrationState::record_satisfiability`]), because a transaction it depends
+    /// on can never mine. Which dependency, and what killed it, is read off that dependency's own
+    /// mark.
+    Inherited,
+}
+
+impl AsRef<str> for UnsatisfiableKind {
+    /// The stable lowercase wire name of this kind, as a store persists it. Borrow-free: it
+    /// returns a `&'static str`, so encoding a kind allocates nothing.
+    fn as_ref(&self) -> &str {
+        match self {
+            UnsatisfiableKind::InputsSpent => "inputs_spent",
+            UnsatisfiableKind::InputsInvalidated => "inputs_invalidated",
+            UnsatisfiableKind::AnchorInvalidated => "anchor_invalidated",
+            UnsatisfiableKind::Inherited => "inherited",
+        }
+    }
+}
+
+/// The error returned when a string does not name an [`UnsatisfiableKind`] (its
+/// [`TryFrom<&str>`] impl).
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub struct ParseUnsatisfiableKindError;
+
+impl fmt::Display for ParseUnsatisfiableKindError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.write_str("unrecognized unsatisfiability kind")
+    }
+}
+
+impl TryFrom<&str> for UnsatisfiableKind {
+    type Error = ParseUnsatisfiableKindError;
+
+    /// Parses the lowercase wire name produced by [`AsRef<str>`](AsRef).
+    ///
+    /// The vocabulary is open — this enum is `#[non_exhaustive]` — but an unrecognized name is
+    /// still an ERROR here, and a store surfaces it as corruption exactly as it does an
+    /// unrecognized [`MigrationStatus`] or [`MigrationTxState`] discriminant. A name written by a
+    /// future release is not a value this one can render or reason about, and treating it as
+    /// "unmarked" would resurrect a dead transaction into the prove and broadcast queues, so it is
+    /// reported rather than smoothed over.
+    ///
+    /// [`MigrationStatus`]: crate::engine::MigrationStatus
+    fn try_from(s: &str) -> Result<Self, Self::Error> {
+        Ok(match s {
+            "inputs_spent" => UnsatisfiableKind::InputsSpent,
+            "inputs_invalidated" => UnsatisfiableKind::InputsInvalidated,
+            "anchor_invalidated" => UnsatisfiableKind::AnchorInvalidated,
+            "inherited" => UnsatisfiableKind::Inherited,
+            _ => return Err(ParseUnsatisfiableKindError),
+        })
+    }
+}
+
+/// A store implementation's observation for one queried nullifier, fed to
+/// [`classify_input_observations`].
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub enum InputObservation {
+    /// The note is known and unspent.
+    Unspent,
+    /// The note is known and the wallet has seen it spent in a mined transaction.
+    SeenSpent,
+    /// The note is known, its creating transaction is unmined, and that transaction's anchor was
+    /// permanently invalidated (settled per the caller's policy). Carries the dead anchor root.
+    Invalidated([u8; 32]),
+    /// No note with this nullifier has been scanned, or its unmined creator is still viable.
+    Unknown,
+}
+
+/// Fold per-nullifier observations, plus the transaction-level expiry judgment, into a
+/// [`StepSatisfiability`], encoding the answer precedence: `InputsSpent` over
+/// `InputsInvalidated` (both are permanent input-level deaths, and a recognized spend is the
+/// definitive on-chain fact) over `Expired` (expiry's remedy — a rebuild, fresh proof included —
+/// cannot cure missing inputs, so the input-level causes must dominate it) over
+/// `NotYetSatisfiable` (anything definite dominates "try later") over `Satisfiable`.
+///
+/// Pure: an implementation supplies only its data access and composes with this, so the
+/// classification is defined once. `AnchorInvalidated` is a transaction-level observation on
+/// broadcast-unmined transactions, answered by the implementation directly rather than through
+/// this per-input fold, and takes the precedence position below `Expired`. The observation set
+/// must come from the transaction's [`spend_nullifiers`] cache, which is non-empty for every
+/// validly committed transaction; implementations surface their corruption error for an empty
+/// cache on a non-mined transaction rather than calling this with no observations (an empty
+/// cache means a backfill-exempted mined row re-entered the watched set after a chain rewind, or
+/// deeper corruption — never vacuous satisfiability).
+///
+/// [`spend_nullifiers`]: crate::engine::MigrationTransaction::spend_nullifiers
+pub fn classify_input_observations(
+    as_of_height: BlockHeight,
+    expired: bool,
+    observations: &[([u8; 32], InputObservation)],
+) -> StepSatisfiability {
+    let spent: Vec<[u8; 32]> = observations
+        .iter()
+        .filter(|(_, o)| matches!(o, InputObservation::SeenSpent))
+        .map(|(nf, _)| *nf)
+        .collect();
+    if !spent.is_empty() {
+        return StepSatisfiability::Unsatisfiable {
+            cause: UnsatisfiableCause::InputsSpent { nullifiers: spent },
+            as_of_height,
+        };
+    }
+    if let Some(anchor) = observations.iter().find_map(|(_, o)| match o {
+        InputObservation::Invalidated(a) => Some(*a),
+        _ => None,
+    }) {
+        return StepSatisfiability::Unsatisfiable {
+            cause: UnsatisfiableCause::InputsInvalidated { anchor },
+            as_of_height,
+        };
+    }
+    if expired {
+        return StepSatisfiability::Unsatisfiable {
+            cause: UnsatisfiableCause::Expired,
+            as_of_height,
+        };
+    }
+    if observations
+        .iter()
+        .any(|(_, o)| matches!(o, InputObservation::Unknown))
+    {
+        return StepSatisfiability::NotYetSatisfiable { as_of_height };
+    }
+    StepSatisfiability::Satisfiable { as_of_height }
+}
+
+/// Drive-level policy for [`advance_migration`]: a struct, so future knobs join without signature
+/// churn.
+#[derive(Clone, Copy, Debug)]
+pub struct AdvanceConfig {
+    reorg_settle_depth: ReorgSettleDepth,
+}
+
+impl AdvanceConfig {
+    /// Drive-level policy under the caller's reorg-settlement policy.
+    ///
+    /// The settle depth is explicitly required — no `Default` — because the right value tracks the
+    /// chain's block spacing, and keeping it current is the caller's responsibility, not this
+    /// crate's.
+    pub const fn new(reorg_settle_depth: ReorgSettleDepth) -> Self {
+        Self { reorg_settle_depth }
+    }
+
+    /// The caller's reorg-settlement policy; see [`ReorgSettleDepth`].
+    pub const fn reorg_settle_depth(&self) -> ReorgSettleDepth {
+        self.reorg_settle_depth
+    }
+}
+
+/// The two heights a migration's dueness is judged against: what the wallet's chain data SUPPORTS,
+/// and where the tip has probably reached.
+///
+/// Both carry this crate's TARGET convention — `chain_tip + 1`, the height of the next block a
+/// transaction could be mined in — so both are directly comparable with an
+/// [`expiry_height`](crate::engine::MigrationTransaction::expiry_height) or a
+/// [`scheduled_height`](crate::engine::MigrationTransaction::scheduled_height); neither is a raw tip.
+///
+/// # Why two
+///
+/// ZIP 318 separates a wallet's SYNC wake-ups from its BROADCAST wake-ups, so a wallet must be able
+/// to decide what is due WITHOUT contacting lightwalletd first (only the Tor-guarded submission
+/// follows the decision). At that moment it holds exactly two things: the frontier its own database
+/// has fully scanned to — an observation — and a wall-clock ESTIMATE of where the chain tip has
+/// reached since. Schedule dueness is unknowable from the frontier alone (a wallet that has not
+/// synced for a week would consider nothing due), so the estimate must be usable; but every
+/// judgment that PERSISTS a verdict or DESTROYS work must rest on chain data, because the state
+/// machine cannot tell a guess from an observation.
+///
+/// # The invariant
+///
+/// No transition that persists a verdict or destroys work may depend on chain state above the
+/// SCANNED frontier. The estimate may only RE-ORDER the service of steps that are already legal, or
+/// protectively WITHHOLD them. Concretely:
+///
+/// | Judged at [`effective`](Self::effective) | Judged at [`scanned`](Self::scanned) |
+/// |---|---|
+/// | schedule dueness in the prove and broadcast queues | expiry as a DECISION: the kernel's dead set, [`MigrationState::expired_transactions`] and [`AdvanceStep::Rebuild`] eligibility, the drain-time [`Replan`](crate::state::AdvanceStep::Replan) gate |
+/// | the doomed-broadcast withhold — a transfer whose expiry has probably passed is not offered, and [`transaction_statuses`](MigrationState::transaction_statuses) reports [`Blocker::ExpiryImminent`](crate::state::Blocker::ExpiryImminent); protective and reversible, since the scanned path still owns the verdict | the dependency closure of [`MigrationState::record_satisfiability`], which writes `Inherited` marks into the store |
+/// | the prove-side expiry skip: proving a transfer that has probably lapsed is wasted work, and skipping it is reversible | anchor-boundary settledness, since an estimate cannot conjure a commitment-tree checkpoint |
+/// | | the [`Expired`](crate::state::Blocker::Expired) blocker in `transaction_statuses`, a rendered determination |
+///
+/// The clamp in [`new`](Self::new) guarantees `effective >= scanned`, so the transposition
+/// hazard — an estimate reaching a destructive check — is unrepresentable by construction rather
+/// than by call-site discipline.
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub struct DuenessTargets {
+    scanned: BlockHeight,
+    effective: BlockHeight,
+}
+
+impl DuenessTargets {
+    /// Targets from the wallet's own scanned frontier and its estimate of the chain tip, both as
+    /// TARGETS (`height + 1`, the next block a transaction could mine in — not the heights
+    /// themselves).
+    ///
+    /// `estimated` is clamped up to `scanned`: an estimate that lags the wallet's own observations
+    /// is simply less informed than they are, and letting it lower the effective target would
+    /// withhold work the chain data already justifies.
+    pub fn new(scanned: BlockHeight, estimated: BlockHeight) -> Self {
+        Self {
+            scanned,
+            effective: core::cmp::max(scanned, estimated),
+        }
+    }
+
+    /// The degenerate pair, for a caller whose estimate IS its chain view — a server that follows
+    /// the chain continuously, or a test.
+    ///
+    /// Every judgment is then made at the same height, which is exactly the single-target behavior
+    /// that preceded this type.
+    pub const fn at(target: BlockHeight) -> Self {
+        Self {
+            scanned: target,
+            effective: target,
+        }
+    }
+
+    /// The target backed by CHAIN DATA (the wallet's fully-scanned frontier, plus one): the height
+    /// every persisted verdict and every destructive judgment is made at.
+    pub const fn scanned(&self) -> BlockHeight {
+        self.scanned
+    }
+
+    /// The target the SCHEDULE is served at (the estimated chain tip, plus one; never below
+    /// [`scanned`](Self::scanned)): the height dueness is judged at, and the height a doomed
+    /// broadcast is withheld from.
+    pub const fn effective(&self) -> BlockHeight {
+        self.effective
+    }
+}
+
+/// Decide the next step to advance a committed migration and VERIFY it against the store before
+/// surfacing it: the entry point a consuming application drives a migration with, against the
+/// caller's [`DuenessTargets`] — its scanned frontier and its estimate of where the chain tip has
+/// reached, both as targets (`height + 1`).
+///
+/// A single call plans, verifies, records, and persists:
+///
+/// 0. It adjudicates any standing BROADCAST-FAILURE REPORT
+///    ([`MigrationState::report_broadcast_failure`]) — a node rejected a broadcast, and the
+///    rejection is another observer's testimony until this wallet has scanned the chain state it
+///    rests on. While the oracle still answers from below the reported tip the call surfaces
+///    [`AdvanceStep::Reevaluate`] and does nothing else; once it answers from at or above that
+///    tip the report is discharged, either recording the ordinary evidence-backed mark or
+///    finding the rejection transient and returning the transaction to the broadcast queue in
+///    this same call.
+/// 1. It asks the planning kernel what to do next, decided purely from the persisted state.
+/// 2. It puts the transaction the kernel names to the store's satisfiability oracle
+///    ([`PoolMigrationRead::check_step_satisfiability`]) — the wallet's live view of whether that
+///    pre-signed artifact can still execute.
+/// 3. A candidate the wallet reports PERMANENTLY obstructed (its inputs seen spent elsewhere, or
+///    unable to ever exist) is recorded through [`MigrationState::record_satisfiability`], and the
+///    kernel is asked again — the recorded marks, and everything stranded behind them, now
+///    filtered out of its queues. A candidate the wallet cannot YET vouch for is set aside for the
+///    rest of this call (no mark: the honest reading of "retry after further sync"), so the
+///    migration's other work still surfaces, and the kernel is asked again.
+/// 4. The step that survives verification is returned, with every determination made along the way
+///    already written back to the store.
+///
+/// So a consumer never spends proving or broadcast work on a transaction whose inputs are gone,
+/// and a migration whose plan has been undercut — most often by an ordinary wallet spend consuming
+/// notes it had allocated — surfaces [`AdvanceStep::Replan`] instead of retrying a dead
+/// transaction forever.
+///
+/// # The drive loop
+///
+/// One call returns ONE step. The consumer performs that step's I/O, records the outcome in the
+/// state, persists it, and calls this again; until the state records the step's completion, the
+/// same step is offered again. The steps map onto the crate's operations as follows:
+///
+/// - [`AdvanceStep::Prove`]: install the transaction's deferred anchor and witnesses and store the
+///   proven PCZT — [`prove_transfer`] / [`prove_preparation`], which also record the
+///   `Signed -> Proved` transition. Proving needs a SYNCED wallet and mutable access to its
+///   commitment trees, but only the account's viewing key. The step carries the transaction's kind
+///   so the consumer can tell, without a lookup, whether the broadcast follows in the same session
+///   (a preparation) or in its own later one (a transfer).
+/// - [`AdvanceStep::Broadcast`]: submit the stored proven transaction to the network, then record
+///   it with [`MigrationState::mark_broadcast`]. Its mining is later detected through the
+///   consumer's own chain view and recorded with [`MigrationState::mark_mined`], which is what
+///   unblocks the transactions depending on it.
+/// - [`AdvanceStep::Rebuild`]: construct and sign a replacement for an expired transfer —
+///   [`rebuild_expired_transfer`] or its unsigned (external-signer) variant. The only step that
+///   needs the account's SPEND AUTHORITY.
+/// - [`AdvanceStep::Replan`]: the migration itself needs replacing — mark it superseded
+///   ([`MigrationState::mark_superseded`]), persist, and re-plan the remaining balance through the
+///   ordinary planning flow. Surfaced when the unsatisfiable share of planned transfer value
+///   strictly exceeds the committed [`ReplanThreshold`], and — once no live work remains at all —
+///   when dead value would otherwise be stranded.
+/// - [`AdvanceStep::Reevaluate`]: SYNC the wallet to at least the tip reported alongside a
+///   rejected broadcast, then call this again. Surfaced ahead of every step but
+///   [`Complete`](AdvanceStep::Complete), and unconditionally: a rejection means another observer
+///   saw chain state this wallet has not, and same-seed activity invalidates the whole store
+///   view rather than one transaction's answer.
+/// - [`AdvanceStep::Waiting`]: nothing is actionable at this height. Consult
+///   [`MigrationState::transaction_statuses`] for what each transaction is blocked on, and register
+///   the heights at which to wake and re-check: [`MigrationState::sync_wakeup_schedule`] for the
+///   proving wake-ups, plus each transaction's own scheduled broadcast height.
+/// - [`AdvanceStep::Complete`]: the migration is terminal (every transaction mined, or the
+///   migration failed/cancelled); nothing will ever be actionable again, so stop polling.
+///
+/// A transaction that can never mine is never named: those marked unsatisfiable, those expired
+/// without mining, and everything stranded behind either are excluded from the prove, broadcast,
+/// and rebuild queues alike. And a DUE BROADCAST is named ahead of any proving work, which is what
+/// makes a broadcast-only waking session possible (see the ZIP 318 note below).
+///
+/// # The dueness contract
+///
+/// `targets` carries two heights because this call makes two KINDS of judgment, and only one of
+/// them may rest on a guess (see [`DuenessTargets`] for the classification table and the ZIP 318
+/// session separation that forces the split). The invariant this function upholds:
+///
+/// > No transition that persists a verdict or destroys work depends on chain state above the
+/// > SCANNED target. The estimate may only re-order the service of steps that are already legal,
+/// > or protectively withhold them.
+///
+/// So the judgments that are DESTRUCTIVE if made too high are all anchored at
+/// [`scanned`](DuenessTargets::scanned):
+///
+/// - EXPIRY as a determination is `expiry_height < scanned`, so an overshooting estimate can never
+///   condemn a live pre-signed transfer as expired — a remedy that requires signing it ANEW.
+/// - Expired-and-unmined transactions SEED the kernel's dead set, so the same overshoot cannot
+///   strand every dependent behind them.
+/// - The durable dependency closure ([`MigrationState::record_satisfiability`]) stamps inherited
+///   marks off those seeds, writing them into the store.
+/// - The drain-time [`Replan`](AdvanceStep::Replan) fires once every unmined transaction is dead,
+///   and the consumer's contracted response to it — [`MigrationState::mark_superseded`] — is
+///   terminal.
+/// - [`Rebuild`](AdvanceStep::Rebuild) eligibility is judged the same way.
+///
+/// while the SCHEDULE is served at [`effective`](DuenessTargets::effective), so a wallet woken by
+/// wall-clock estimate submits a transfer that is due without first synchronizing. The one further
+/// thing the estimate does is WITHHOLD: a proven transfer whose expiry has probably passed
+/// (`scanned <= expiry_height < effective`) is not offered for broadcast or proving, because
+/// submitting it would be refused by the node and proving it would be wasted work. That withhold
+/// records nothing and reverses itself as the wallet's scan catches up, whereupon the scanned path
+/// makes the actual determination. [`MigrationState::transaction_statuses`] renders such a
+/// transaction as [`Blocker::ExpiryImminent`](crate::state::Blocker::ExpiryImminent), neither
+/// `ready` nor carrying an action, so a status-driven consumer never wakes a broadcast session
+/// this function would refuse.
+///
+/// # What a call costs
+///
+/// Checking is LAZY UNTIL FIRST DISCOVERY and BROAD AT DISCOVERY, so a healthy migration pays
+/// almost nothing:
+///
+/// - the in-flight sweep — one query per broadcast-but-unmined transaction, judging whether a
+///   settled reorg has permanently invalidated the anchor it was proven against. The broadcast
+///   schedule keeps that set tiny (typically none to two transactions sit between broadcast and
+///   mining), and it is the only check for transactions the kernel will never name as a step;
+/// - one query for the candidate the kernel names — and none at all when the kernel answers
+///   [`Waiting`](AdvanceStep::Waiting), [`Complete`](AdvanceStep::Complete), or
+///   [`Replan`](AdvanceStep::Replan), which name no transaction.
+///
+/// A discovery costs more, deliberately: it also checks every pending transaction whose
+/// dependencies have mined. That is what makes the migration's remaining value re-assessed in the
+/// call that finds the first dead transaction — a send-max that consumed the funding notes trips
+/// the replan threshold immediately, rather than one transfer at a time over the weeks a
+/// privacy-preserving broadcast schedule spans.
+///
+/// # Persistence, and what the consumer owns
+///
+/// This function owns the persistence of the determinations it makes: whenever a check recorded
+/// anything, the state is written back with [`PoolMigrationWrite::replace_migration`] BEFORE the
+/// step is returned, so the store already agrees with the returned `state` when the consumer acts.
+/// Nothing is written when nothing was discovered.
+///
+/// On an `Err`, treat the passed `state` as UNTRUSTED and re-read it with
+/// [`get_migration`](PoolMigrationRead::get_migration) before driving again. A failing store
+/// leaves the two views free to disagree in one direction: the store never holds a determination
+/// the returned state lacks, but the reverse can happen — a write that failed, or in-flight-sweep
+/// marks left unpersisted by a later oracle failure, live only in the in-memory state. Discarding
+/// it costs nothing, because marks are re-derivable: the next call asks the same questions again.
+///
+/// Everything else remains the consumer's: it performs the returned step's I/O and records the
+/// outcome through the ordinary mutators ([`MigrationState::mark_broadcast`],
+/// [`MigrationState::mark_mined`], the prove and rebuild functions), persisting afterwards as
+/// usual, then calls this again. In particular, the ZIP 318 separation of SYNC wake-ups from
+/// BROADCAST wake-ups is the consumer's runtime policy. Broadcasting a stored proven transaction
+/// needs no sync at all, while proving is inherently sync-bound, so surfacing every due broadcast
+/// first is what lets a wallet wake, submit, and end the session without initiating sync
+/// operations. But this function itself has no notion of a session — once every due broadcast is
+/// dispatched it offers proving work in the same loop — so a wallet honoring the separation stops
+/// driving the migration after broadcasting and leaves the offered proving work to its next sync
+/// wake-up.
+///
+/// [`prove_transfer`]: crate::engine::prove_transfer
+/// [`prove_preparation`]: crate::engine::prove_preparation
+/// [`rebuild_expired_transfer`]: crate::engine::rebuild_expired_transfer
+pub fn advance_migration<St: PoolMigrationWrite>(
+    store: &mut St,
+    state: &mut MigrationState,
+    targets: DuenessTargets,
+    config: &AdvanceConfig,
+) -> Result<AdvanceStep, St::Error> {
+    // Whether any determination has been recorded, and so whether the state must be written back
+    // before a step is surfaced.
+    let mut dirty = false;
+    // This call's deferrals: candidates the wallet answered `NotYetSatisfiable` for. They carry no
+    // mark — the answer is "retry after further sync", not an obstruction — so the list lives only
+    // for this call.
+    let mut set_aside: Vec<MigrationTransferId> = Vec::new();
+
+    // The IN-FLIGHT SWEEP. A broadcast-but-unmined transaction is never named as a step, so no
+    // candidate check would ever reach it; its liveness question is whether the anchor it was
+    // PROVEN against still exists on the chain, which only a settled reorg answers no to. That is
+    // `AnchorInvalidated`, and it is the ONLY cause this sweep records:
+    //
+    // - `InputsSpent` about an in-flight transaction is most likely the transaction ITSELF, seen
+    //   spending its own inputs in the block that mined it while this state still says
+    //   `Broadcast` (the consumer's `mark_mined` lags the wallet's scan). Marking on it would
+    //   kill a transaction that is mining.
+    // - `InputsInvalidated` and `Expired` are DELIBERATELY deferred rather than acted on here.
+    //   Their loss is bounded: a producer inside the migration reaches its victim through the
+    //   dependency closure, and a victim of an EXTERNAL producer self-heals at its own expiry,
+    //   when the kernel's derived expiry makes it a dead-set seed. The pending-side channel — a
+    //   candidate check, or the broaden sweep after any discovery — is where an input-level cause
+    //   is a determination.
+    //
+    // One consequence to keep in view: the answer precedence puts `InputsSpent` above the
+    // anchor-level cause, so a store that observes BOTH a spent input and a dead anchor reports
+    // `InputsSpent`, which this sweep drops. That transaction's death then surfaces at its expiry
+    // rather than now — later, never wrong.
+    let mut findings: Vec<(MigrationTransferId, StepSatisfiability)> = Vec::new();
+    for tx in state.transactions() {
+        // In flight, and not already marked: exactly the transactions this sweep is about.
+        if matches!(tx.state(), MigrationTxState::Broadcast { .. })
+            && tx.unsatisfiable_at().is_none()
+        {
+            let answer = store.check_step_satisfiability(tx, config.reorg_settle_depth())?;
+            if matches!(
+                &answer,
+                StepSatisfiability::Unsatisfiable {
+                    cause: UnsatisfiableCause::AnchorInvalidated,
+                    ..
+                }
+            ) {
+                findings.push((tx.id(), answer));
+            }
+        }
+    }
+    if !findings.is_empty() {
+        state.record_satisfiability(targets, &findings);
+        dirty = true;
+    }
+
+    // ADJUDICATING THE BROADCAST-FAILURE REPORTS. A report says a node refused a broadcast at a
+    // tip this wallet may not have reached; it is testimony, so nothing is concluded from it
+    // until the wallet's own answers rest at or above that tip. One query per reported
+    // transaction, and none at all for the overwhelmingly common case of no reports.
+    //
+    // Skipped entirely for a TERMINAL migration: it is never driven further, so there is no
+    // broadcast to withhold and no benefit in holding the loop at `Reevaluate` instead of the
+    // `Complete` the kernel is about to return.
+    let mut reevaluation_pending = false;
+    if !state.is_terminal() {
+        let mut adjudicated: Vec<MigrationTransferId> = Vec::new();
+        let mut verdicts: Vec<(MigrationTransferId, StepSatisfiability)> = Vec::new();
+        for tx in state.transactions() {
+            let Some(reported_tip) = tx.broadcast_failure_at() else {
+                continue;
+            };
+            let answer = store.check_step_satisfiability(tx, config.reorg_settle_depth())?;
+            if answer_as_of_height(&answer) < reported_tip {
+                // The wallet has not scanned every block that could hold the spend behind the
+                // rejection, so neither answer it could give would be about the right chain.
+                reevaluation_pending = true;
+                continue;
+            }
+            // At or above the reported tip the question is settled, whichever way it fell: a
+            // marking cause records the ordinary evidence-backed determination, and anything
+            // else means the rejection was transient (a mempool conflict, a node-local policy)
+            // and the transaction returns to the broadcast queue in this same call. Either way
+            // the testimony has been discharged and must not withhold the transaction again.
+            adjudicated.push(tx.id());
+            if records_a_determination(&answer) {
+                verdicts.push((tx.id(), answer));
+            }
+        }
+        if !verdicts.is_empty() {
+            // A DISCOVERY here broadens exactly as one at a candidate check does: what killed
+            // this transaction is a fact about the migration's environment — same-seed spending
+            // typically kills several transfers at once — so the whole damage is assessed in the
+            // call that finds the first of it, and the replan threshold trips here rather than
+            // one broadcast attempt at a time.
+            broaden_after_discovery(store, state, &mut verdicts, config.reorg_settle_depth())?;
+            state.record_satisfiability(targets, &verdicts);
+        }
+        for id in &adjudicated {
+            state.clear_broadcast_failure(*id);
+        }
+        dirty |= !adjudicated.is_empty();
+    }
+    if reevaluation_pending {
+        // Unconditional: a rejection means some other observer saw chain state this wallet has
+        // not, and same-seed activity invalidates the whole store view rather than one
+        // transaction's answer. Whatever was adjudicated above is still persisted first.
+        if dirty {
+            store.replace_migration(state)?;
+        }
+        return Ok(AdvanceStep::Reevaluate);
+    }
+
+    // Plan, verify, record, plan again. Every iteration either breaks, grows `set_aside`, or
+    // strictly grows the marked set — strictly, because the kernel never names an already-marked
+    // transaction (they are in its dead set), so a recorded discovery marks at least the candidate
+    // that produced it. Both are bounded by the transaction count, so the loop terminates.
+    let step = loop {
+        let step = state.next_step(targets, &set_aside);
+        let candidate = match step {
+            AdvanceStep::Prove { id, .. }
+            | AdvanceStep::Broadcast { id }
+            | AdvanceStep::Rebuild { id } => id,
+            // These name no transaction, so there is nothing to verify. `Reevaluate` is decided
+            // above, by this function, and the kernel never returns it — but it is a step like
+            // any other here, and matching it explicitly keeps that fact from resting on the
+            // absence of an arm.
+            AdvanceStep::Waiting
+            | AdvanceStep::Complete
+            | AdvanceStep::Replan
+            | AdvanceStep::Reevaluate => break step,
+        };
+        let Some(tx) = state.transactions().iter().find(|t| t.id() == candidate) else {
+            // The kernel names its candidates out of `state.transactions()`, so an id that is not
+            // there means the state is corrupt. Nothing can be verified about it and nothing may
+            // be surfaced unverified, so the call reports `Waiting` — the safe degradation, and a
+            // BREAK rather than another iteration, which leaves no way for a corrupt state to
+            // spin the loop. Anything already recorded is still persisted below.
+            break AdvanceStep::Waiting;
+        };
+        let answer = store.check_step_satisfiability(tx, config.reorg_settle_depth())?;
+        match answer {
+            StepSatisfiability::Satisfiable { .. } => break step,
+            // Not an obstruction: the wallet cannot yet vouch for the inputs (their source is
+            // unmined, or mined but unscanned). Take the candidate out of this call's queues so
+            // the migration's other work still surfaces, and leave no mark — a false mark would
+            // strand live value behind an observation only a reorg can clear.
+            StepSatisfiability::NotYetSatisfiable { .. } => set_aside.push(candidate),
+            answer @ StepSatisfiability::Unsatisfiable { .. } => {
+                if !records_a_determination(&answer) {
+                    // `Expired`: never overrides the kernel, which derives expiry itself from the
+                    // same `expiry_height`. The two can disagree only under caller/store height
+                    // skew, and there the kernel's scanned target governs, so the step stands
+                    // exactly as planned (for a transfer, the rebuild that is expiry's remedy).
+                    break step;
+                }
+                // A DISCOVERY, so BROADEN, then record as ONE batch so the durable closure runs
+                // once over the whole discovery.
+                let mut batch = vec![(candidate, answer)];
+                broaden_after_discovery(store, state, &mut batch, config.reorg_settle_depth())?;
+                state.record_satisfiability(targets, &batch);
+                dirty = true;
+            }
+        }
+    };
+
+    // A determination is durable by the time the step it produced is surfaced: the consumer acts
+    // on that step and persists its own outcome, and must not be able to lose the marks this call
+    // discovered along the way.
+    if dirty {
+        store.replace_migration(state)?;
+    }
+    Ok(step)
+}
+
+/// Whether a satisfiability answer is a DISCOVERY for [`advance_migration`] to record and broaden
+/// on: an obstruction whose cause marks ([`UnsatisfiableCause::marks`], the single definition of
+/// that decision). The satisfiable and not-yet answers are not determinations at all.
+fn records_a_determination(answer: &StepSatisfiability) -> bool {
+    matches!(answer, StepSatisfiability::Unsatisfiable { cause, .. } if cause.marks())
+}
+
+/// The fully-scanned height a satisfiability answer rests on, whichever answer it is: the height
+/// that decides whether the answer is ABOUT the chain state a question was raised at.
+fn answer_as_of_height(answer: &StepSatisfiability) -> BlockHeight {
+    match answer {
+        StepSatisfiability::Satisfiable { as_of_height }
+        | StepSatisfiability::NotYetSatisfiable { as_of_height }
+        | StepSatisfiability::Unsatisfiable { as_of_height, .. } => *as_of_height,
+    }
+}
+
+/// Grows `batch` — the determinations [`advance_migration`] has already discovered this call —
+/// with every OTHER determination the store will report right now, so a whole discovery is
+/// assessed at once.
+///
+/// What killed one transaction is a fact about the migration's ENVIRONMENT, not about that
+/// transaction alone: the ordinary cause is a wallet spend consuming notes the migration had
+/// allocated, which typically kills several transfers together. Finding the rest of the damage in
+/// the same call is what lets the replan threshold trip immediately rather than one transfer at a
+/// time over the weeks a privacy-preserving broadcast schedule spans.
+///
+/// Every PENDING transaction whose dependencies have mined is checked, and nothing else: a
+/// transaction already in `batch` has been answered, mined ones are final, in-flight ones belong
+/// to the drive API's own in-flight sweep, already-marked ones would learn nothing, and one whose
+/// dependencies are still unmined could only answer "not yet" (its death, if any, arrives through
+/// the durable dependency closure).
+fn broaden_after_discovery<St: PoolMigrationRead>(
+    store: &St,
+    state: &MigrationState,
+    batch: &mut Vec<(MigrationTransferId, StepSatisfiability)>,
+    settle: ReorgSettleDepth,
+) -> Result<(), St::Error> {
+    for tx in state.transactions() {
+        // Unanswered this call, pending, unmarked, and with its dependencies mined: exactly the
+        // transactions this sweep is about.
+        if !batch.iter().any(|(id, _)| *id == tx.id())
+            && !matches!(
+                tx.state(),
+                MigrationTxState::Broadcast { .. } | MigrationTxState::Mined { .. }
+            )
+            && tx.unsatisfiable_at().is_none()
+            && state.deps_mined(tx.depends_on())
+        {
+            let answer = store.check_step_satisfiability(tx, settle)?;
+            if records_a_determination(&answer) {
+                batch.push((tx.id(), answer));
+            }
+        }
+    }
+    Ok(())
+}
+
+/// The drive API's tests, over a minimal in-crate store with a configurable answer set: what
+/// [`advance_migration`] surfaces, what it records, how much it asks, and what it persists.
+#[cfg(test)]
+mod advance_tests {
+    use super::*;
+    use alloc::collections::BTreeMap;
+    use core::cell::Cell;
+    use zcash_protocol::{TxId, value::Zatoshis};
+
+    use crate::denomination::DenominationPlan;
+    use crate::engine::{MigrationStatus, MigrationTransaction, MigrationTxKind};
+    use crate::preparation::PreparationPlan;
+
+    /// A store that answers satisfiability from a fixed table — anything absent is satisfiable at
+    /// `as_of_height`, the healthy default — and counts what the drive loop asks of it.
+    struct TestStore {
+        stored: Option<MigrationState>,
+        satisfiability: BTreeMap<MigrationTransferId, StepSatisfiability>,
+        as_of_height: BlockHeight,
+        queries: Cell<usize>,
+        replaced: Cell<usize>,
+    }
+
+    impl TestStore {
+        fn new(
+            as_of_height: u32,
+            answers: impl IntoIterator<Item = (MigrationTransferId, StepSatisfiability)>,
+        ) -> Self {
+            TestStore {
+                stored: None,
+                satisfiability: answers.into_iter().collect(),
+                as_of_height: BlockHeight::from_u32(as_of_height),
+                queries: Cell::new(0),
+                replaced: Cell::new(0),
+            }
+        }
+
+        /// Replace one transaction's answer, standing in for the wallet's view changing between
+        /// calls (a dependency scanned, a reorg settled).
+        fn set_answer(&mut self, id: MigrationTransferId, answer: StepSatisfiability) {
+            self.satisfiability.insert(id, answer);
+        }
+    }
+
+    impl PoolMigrationRead for TestStore {
+        type Error = core::convert::Infallible;
+
+        fn get_migration(&self) -> Result<Option<MigrationState>, Self::Error> {
+            Ok(self.stored.clone())
+        }
+
+        fn check_step_satisfiability(
+            &self,
+            tx: &MigrationTransaction,
+            _settle: ReorgSettleDepth,
+        ) -> Result<StepSatisfiability, Self::Error> {
+            self.queries.set(self.queries.get() + 1);
+            Ok(self.satisfiability.get(&tx.id()).cloned().unwrap_or(
+                StepSatisfiability::Satisfiable {
+                    as_of_height: self.as_of_height,
+                },
+            ))
+        }
+    }
+
+    impl PoolMigrationWrite for TestStore {
+        fn replace_migration(&mut self, state: &MigrationState) -> Result<(), Self::Error> {
+            self.replaced.set(self.replaced.get() + 1);
+            self.stored = Some(state.clone());
+            Ok(())
+        }
+
+        fn update_transaction(
+            &mut self,
+            id: MigrationTransferId,
+            state: MigrationTxState,
+        ) -> Result<(), Self::Error> {
+            if let Some(stored) = &mut self.stored
+                && let Some(tx) = stored.transactions.iter_mut().find(|t| t.id == id)
+            {
+                tx.state = state;
+            }
+            Ok(())
+        }
+    }
+
+    // The state constructors mirror `state.rs`'s test helpers, so a drive-loop test and a kernel
+    // test describe the same migration the same way.
+
+    fn tx(id: u32, kind: MigrationTxKind, state: MigrationTxState) -> MigrationTransaction {
+        MigrationTransaction {
+            id: MigrationTransferId(id),
+            kind,
+            pczt: Vec::new(),
+            depends_on: Vec::new(),
+            scheduled_height: BlockHeight::from_u32(0),
+            expiry_height: BlockHeight::from_u32(0),
+            anchor_boundary: None,
+            state,
+            lock_owner: None,
+            unsatisfiable: None,
+            spend_nullifiers: Vec::new(),
+            broadcast_failure_at: None,
+        }
+    }
+
+    fn prep(layer: usize, index: usize) -> MigrationTxKind {
+        MigrationTxKind::Preparation { layer, index }
+    }
+
+    fn transfer(crossing: usize) -> MigrationTxKind {
+        MigrationTxKind::Transfer { crossing }
+    }
+
+    fn mined(height: u32) -> MigrationTxState {
+        MigrationTxState::Mined {
+            txid: TxId::from_bytes([0; 32]),
+            height: BlockHeight::from_u32(height),
+        }
+    }
+
+    fn broadcast() -> MigrationTxState {
+        MigrationTxState::Broadcast {
+            txid: TxId::from_bytes([1; 32]),
+        }
+    }
+
+    /// A transfer with the given anchor boundary and scheduled broadcast height, in the given
+    /// lifecycle state (never expiring, like `tx`).
+    fn scheduled_transfer(
+        id: u32,
+        crossing: usize,
+        anchor: u32,
+        broadcast: u32,
+        state: MigrationTxState,
+    ) -> MigrationTransaction {
+        let mut t = tx(id, transfer(crossing), state);
+        t.anchor_boundary = Some(BlockHeight::from_u32(anchor));
+        t.scheduled_height = BlockHeight::from_u32(broadcast);
+        t
+    }
+
+    /// A state whose denomination plan carries the given crossing values, so each transfer's
+    /// contribution to the replan threshold is exactly the value at its crossing index.
+    fn state_with_crossings(
+        crossings: &[u64],
+        transactions: Vec<MigrationTransaction>,
+    ) -> MigrationState {
+        let zats = |v: u64| Zatoshis::from_u64(v).expect("test values are valid");
+        let total = zats(crossings.iter().sum());
+        MigrationState {
+            status: MigrationStatus::Committed,
+            denominations: DenominationPlan::from_stored_parts(
+                crossings.iter().copied().map(zats).collect(),
+                zats(15_000),
+                None,
+                Zatoshis::ZERO,
+                total,
+                total,
+            )
+            .expect("a consistent stored plan reconstructs"),
+            preparation: PreparationPlan::from_parts(Vec::new(), Vec::new()),
+            transactions,
+            anchor_bucket_interval: crate::scheduling::AnchorBucketInterval::ZIP_318,
+            replan_threshold: ReplanThreshold::DEFAULT,
+        }
+    }
+
+    fn spent(as_of_height: u32) -> StepSatisfiability {
+        StepSatisfiability::Unsatisfiable {
+            cause: UnsatisfiableCause::InputsSpent {
+                nullifiers: vec![[9; 32]],
+            },
+            as_of_height: BlockHeight::from_u32(as_of_height),
+        }
+    }
+
+    fn config() -> AdvanceConfig {
+        AdvanceConfig::new(ReorgSettleDepth::new(10))
+    }
+
+    /// Which half of the store interface fails, for the error-propagation test.
+    #[derive(Clone, Copy, Debug, PartialEq, Eq)]
+    enum Failure {
+        /// The satisfiability oracle fails.
+        Oracle,
+        /// Persisting fails.
+        Write,
+    }
+
+    /// A store whose failing half is chosen up front. `TestStore` is infallible, so error
+    /// propagation needs a store that can actually fail.
+    #[derive(Debug, PartialEq, Eq)]
+    struct StoreFailed;
+
+    struct FailingStore {
+        stored: Option<MigrationState>,
+        satisfiability: BTreeMap<MigrationTransferId, StepSatisfiability>,
+        as_of_height: BlockHeight,
+        fails: Failure,
+        queries: Cell<usize>,
+        replaced: Cell<usize>,
+    }
+
+    impl FailingStore {
+        fn new(
+            fails: Failure,
+            as_of_height: u32,
+            answers: impl IntoIterator<Item = (MigrationTransferId, StepSatisfiability)>,
+        ) -> Self {
+            FailingStore {
+                stored: None,
+                satisfiability: answers.into_iter().collect(),
+                as_of_height: BlockHeight::from_u32(as_of_height),
+                fails,
+                queries: Cell::new(0),
+                replaced: Cell::new(0),
+            }
+        }
+    }
+
+    impl PoolMigrationRead for FailingStore {
+        type Error = StoreFailed;
+
+        fn get_migration(&self) -> Result<Option<MigrationState>, Self::Error> {
+            Ok(self.stored.clone())
+        }
+
+        fn check_step_satisfiability(
+            &self,
+            tx: &MigrationTransaction,
+            _settle: ReorgSettleDepth,
+        ) -> Result<StepSatisfiability, Self::Error> {
+            self.queries.set(self.queries.get() + 1);
+            if self.fails == Failure::Oracle {
+                return Err(StoreFailed);
+            }
+            Ok(self.satisfiability.get(&tx.id()).cloned().unwrap_or(
+                StepSatisfiability::Satisfiable {
+                    as_of_height: self.as_of_height,
+                },
+            ))
+        }
+    }
+
+    impl PoolMigrationWrite for FailingStore {
+        fn replace_migration(&mut self, state: &MigrationState) -> Result<(), Self::Error> {
+            self.replaced.set(self.replaced.get() + 1);
+            if self.fails == Failure::Write {
+                return Err(StoreFailed);
+            }
+            self.stored = Some(state.clone());
+            Ok(())
+        }
+
+        fn update_transaction(
+            &mut self,
+            _id: MigrationTransferId,
+            _state: MigrationTxState,
+        ) -> Result<(), Self::Error> {
+            Err(StoreFailed)
+        }
+    }
+
+    /// The invariant the drive API exists for: a step is surfaced only after the store has
+    /// vouched for it. The kernel plans A's due broadcast; the check discovers A's inputs spent,
+    /// which broadens into a sweep that also finds C dead, and the call surfaces B's prove — never
+    /// A's broadcast — with both discoveries recorded and persisted in ONE write. A's and C's
+    /// value stays below the replan threshold, so the migration keeps draining.
+    #[test]
+    fn advance_never_surfaces_a_dead_candidate_and_broadens_on_discovery() {
+        let mut state = state_with_crossings(
+            &[5_000_000, 90_000_000, 5_000_000],
+            vec![
+                tx(0, prep(0, 0), mined(10)),
+                // A: proved and due at the target height.
+                scheduled_transfer(1, 0, 1440, 1500, MigrationTxState::Proved),
+                // B and C: signed, boundaries long settled, broadcasts weeks out.
+                scheduled_transfer(2, 1, 1440, 90_000, MigrationTxState::Signed),
+                scheduled_transfer(3, 2, 1440, 90_000, MigrationTxState::Signed),
+            ],
+        );
+        let mut store = TestStore::new(
+            1600,
+            [
+                (MigrationTransferId(1), spent(1600)),
+                (MigrationTransferId(3), spent(1600)),
+            ],
+        );
+
+        let step = advance_migration(
+            &mut store,
+            &mut state,
+            DuenessTargets::at(BlockHeight::from_u32(1601)),
+            &config(),
+        )
+        .expect("the store never fails");
+
+        assert!(
+            matches!(step, AdvanceStep::Prove { id, .. } if id == MigrationTransferId(2)),
+            "the live transfer's prove is surfaced, not the dead transfer's broadcast: {step:?}"
+        );
+        let marked = BlockHeight::from_u32(1600);
+        assert_eq!(
+            state.transactions()[1].unsatisfiable_at(),
+            Some(marked),
+            "the dead candidate is marked at the answer's height"
+        );
+        assert_eq!(
+            state.transactions()[3].unsatisfiable_at(),
+            Some(marked),
+            "the broaden sweep marks a transfer whose own broadcast is weeks away"
+        );
+        assert_eq!(
+            state.transactions()[2].unsatisfiable_at(),
+            None,
+            "the satisfiable transfer is untouched"
+        );
+        assert_eq!(
+            store.replaced.get(),
+            1,
+            "one dirty write persisted the marks"
+        );
+        assert_eq!(
+            store.stored.as_ref().expect("persisted").transactions()[3].unsatisfiable_at(),
+            Some(marked),
+            "the store holds the marks by the time the step is surfaced"
+        );
+    }
+
+    /// Broadening is what makes the threshold trip promptly: the marks discovered in one call
+    /// carry the unsatisfiable share past the committed threshold, so that SAME call re-plans and
+    /// returns `Replan` rather than another step.
+    #[test]
+    fn advance_replan_when_threshold_crossed() {
+        let mut state = state_with_crossings(
+            &[30_000_000, 40_000_000, 30_000_000],
+            vec![
+                tx(0, prep(0, 0), mined(10)),
+                scheduled_transfer(1, 0, 1440, 1500, MigrationTxState::Proved),
+                scheduled_transfer(2, 1, 1440, 90_000, MigrationTxState::Signed),
+                scheduled_transfer(3, 2, 1440, 90_000, MigrationTxState::Signed),
+            ],
+        );
+        let mut store = TestStore::new(
+            1600,
+            [
+                (MigrationTransferId(1), spent(1600)),
+                (MigrationTransferId(3), spent(1600)),
+            ],
+        );
+
+        let step = advance_migration(
+            &mut store,
+            &mut state,
+            DuenessTargets::at(BlockHeight::from_u32(1601)),
+            &config(),
+        )
+        .expect("the store never fails");
+
+        assert_eq!(
+            step,
+            AdvanceStep::Replan,
+            "60% of the planned crossing value is unsatisfiable, past the 20% threshold"
+        );
+        assert!(state.replan_required());
+        assert_eq!(store.replaced.get(), 1);
+        let stored = store.stored.as_ref().expect("persisted");
+        assert!(
+            stored.transactions()[1].unsatisfiable_at().is_some()
+                && stored.transactions()[3].unsatisfiable_at().is_some(),
+            "the marks the replan decision rests on are durable before it is surfaced"
+        );
+    }
+
+    /// A `NotYetSatisfiable` answer is a retry, not an obstruction: the candidate is set aside for
+    /// the rest of the call — nothing marked — its sibling surfaces instead, and a FRESH call
+    /// offers the deferred transaction again once the wallet can vouch for it.
+    #[test]
+    fn advance_not_yet_satisfiable_sets_aside_without_marking() {
+        let mut state = state_with_crossings(
+            &[50_000_000, 50_000_000],
+            vec![
+                tx(0, prep(0, 0), mined(10)),
+                scheduled_transfer(1, 0, 1440, 1500, MigrationTxState::Proved),
+                scheduled_transfer(2, 1, 1440, 1500, MigrationTxState::Proved),
+            ],
+        );
+        let mut store = TestStore::new(
+            1600,
+            [(
+                MigrationTransferId(1),
+                StepSatisfiability::NotYetSatisfiable {
+                    as_of_height: BlockHeight::from_u32(1600),
+                },
+            )],
+        );
+
+        let step = advance_migration(
+            &mut store,
+            &mut state,
+            DuenessTargets::at(BlockHeight::from_u32(1601)),
+            &config(),
+        )
+        .expect("the store never fails");
+
+        assert_eq!(
+            step,
+            AdvanceStep::Broadcast {
+                id: MigrationTransferId(2)
+            },
+            "the sibling's due broadcast surfaces while the lagging candidate waits"
+        );
+        assert!(
+            state
+                .transactions()
+                .iter()
+                .all(|t| t.unsatisfiable_at().is_none()),
+            "a not-yet answer marks nothing"
+        );
+        assert_eq!(
+            store.replaced.get(),
+            0,
+            "nothing was recorded, so nothing is written"
+        );
+
+        // The deferral was call-local: once the wallet can vouch for it, the next call offers the
+        // same transaction again.
+        store.set_answer(
+            MigrationTransferId(1),
+            StepSatisfiability::Satisfiable {
+                as_of_height: BlockHeight::from_u32(1700),
+            },
+        );
+        let step = advance_migration(
+            &mut store,
+            &mut state,
+            DuenessTargets::at(BlockHeight::from_u32(1701)),
+            &config(),
+        )
+        .expect("the store never fails");
+        assert_eq!(
+            step,
+            AdvanceStep::Broadcast {
+                id: MigrationTransferId(1)
+            },
+            "set_aside does not persist across calls"
+        );
+        assert_eq!(store.replaced.get(), 0);
+    }
+
+    /// The laziness profile a healthy migration pays: one query for the candidate the kernel
+    /// names, no write at all — and, when nothing is actionable, not even that.
+    #[test]
+    fn advance_is_lazy_when_healthy() {
+        let mut state = state_with_crossings(
+            &[100_000_000],
+            vec![
+                tx(0, prep(0, 0), mined(10)),
+                scheduled_transfer(1, 0, 1440, 1500, MigrationTxState::Proved),
+            ],
+        );
+        let mut store = TestStore::new(1600, []);
+
+        let step = advance_migration(
+            &mut store,
+            &mut state,
+            DuenessTargets::at(BlockHeight::from_u32(1601)),
+            &config(),
+        )
+        .expect("the store never fails");
+
+        assert_eq!(
+            step,
+            AdvanceStep::Broadcast {
+                id: MigrationTransferId(1)
+            }
+        );
+        assert_eq!(store.queries.get(), 1, "exactly the candidate is checked");
+        assert_eq!(store.replaced.get(), 0, "a healthy call writes nothing");
+
+        // Nothing in flight and nothing actionable — the transfer's boundary has yet to settle —
+        // so there is nothing to check at all.
+        let mut waiting = state_with_crossings(
+            &[100_000_000],
+            vec![
+                tx(0, prep(0, 0), mined(10)),
+                scheduled_transfer(1, 0, 1700, 90_000, MigrationTxState::Signed),
+            ],
+        );
+        let mut store = TestStore::new(1600, []);
+        let step = advance_migration(
+            &mut store,
+            &mut waiting,
+            DuenessTargets::at(BlockHeight::from_u32(1601)),
+            &config(),
+        )
+        .expect("the store never fails");
+        assert_eq!(step, AdvanceStep::Waiting);
+        assert_eq!(store.queries.get(), 0, "a waiting call asks nothing");
+        assert_eq!(store.replaced.get(), 0);
+    }
+
+    /// The in-flight sweep is the only check a broadcast-but-unmined transaction ever gets, and it
+    /// records exactly one cause. A settled reorg invalidates a broadcast preparation's anchor;
+    /// its pending dependents die through the closure, carrying the migration past the threshold.
+    /// A sibling in flight whose answer is an INPUT-level cause is left alone: an in-flight
+    /// transaction's inputs are spent by itself, so that answer says nothing about its fate.
+    #[test]
+    fn advance_in_flight_anchor_invalidated_flows_to_replan() {
+        let mut dependent_a = scheduled_transfer(1, 0, 1440, 90_000, MigrationTxState::Signed);
+        dependent_a.depends_on = vec![MigrationTransferId(0)];
+        let mut dependent_b = scheduled_transfer(2, 1, 1440, 90_000, MigrationTxState::Signed);
+        dependent_b.depends_on = vec![MigrationTransferId(0)];
+        let mut state = state_with_crossings(
+            &[50_000_000, 50_000_000, 10_000_000],
+            vec![
+                tx(0, prep(0, 0), broadcast()),
+                dependent_a,
+                dependent_b,
+                // In flight, independent, and answering an input-level cause.
+                scheduled_transfer(3, 2, 1440, 1500, broadcast()),
+            ],
+        );
+        let mut store = TestStore::new(
+            1600,
+            [
+                (
+                    MigrationTransferId(0),
+                    StepSatisfiability::Unsatisfiable {
+                        cause: UnsatisfiableCause::AnchorInvalidated,
+                        as_of_height: BlockHeight::from_u32(1600),
+                    },
+                ),
+                (MigrationTransferId(3), spent(1600)),
+            ],
+        );
+
+        let step = advance_migration(
+            &mut store,
+            &mut state,
+            DuenessTargets::at(BlockHeight::from_u32(1601)),
+            &config(),
+        )
+        .expect("the store never fails");
+
+        assert_eq!(
+            step,
+            AdvanceStep::Replan,
+            "the stranded dependents carry the migration past the threshold"
+        );
+        let marked = BlockHeight::from_u32(1600);
+        assert_eq!(state.transactions()[0].unsatisfiable_at(), Some(marked));
+        assert_eq!(
+            state.transactions()[1].unsatisfiable_at(),
+            Some(marked),
+            "a dependent of the dead preparation inherits its stamp"
+        );
+        assert_eq!(state.transactions()[2].unsatisfiable_at(), Some(marked));
+        assert_eq!(
+            state.transactions()[3].unsatisfiable_at(),
+            None,
+            "an input-level answer about an in-flight transaction is not a determination"
+        );
+        assert_eq!(
+            store.queries.get(),
+            2,
+            "one query per in-flight transaction"
+        );
+        assert_eq!(store.replaced.get(), 1);
+        assert_eq!(
+            store.stored.as_ref().expect("persisted").transactions()[1].unsatisfiable_at(),
+            Some(marked)
+        );
+    }
+
+    /// An `Expired` answer confirms a derivation the kernel already makes from the same
+    /// `expiry_height`, so it never overrides the planned step: the rebuild the kernel surfaced —
+    /// expiry's remedy — stands, and nothing is marked (a mark would make the transfer dead
+    /// rather than rebuildable).
+    #[test]
+    fn advance_expired_answer_never_overrides_the_kernel() {
+        let mut expired = scheduled_transfer(1, 0, 1440, 1500, MigrationTxState::Signed);
+        expired.expiry_height = BlockHeight::from_u32(1550);
+        let mut state =
+            state_with_crossings(&[100_000_000], vec![tx(0, prep(0, 0), mined(10)), expired]);
+        let mut store = TestStore::new(
+            1600,
+            [(
+                MigrationTransferId(1),
+                StepSatisfiability::Unsatisfiable {
+                    cause: UnsatisfiableCause::Expired,
+                    as_of_height: BlockHeight::from_u32(1600),
+                },
+            )],
+        );
+
+        let step = advance_migration(
+            &mut store,
+            &mut state,
+            DuenessTargets::at(BlockHeight::from_u32(1601)),
+            &config(),
+        )
+        .expect("the store never fails");
+
+        assert_eq!(
+            step,
+            AdvanceStep::Rebuild {
+                id: MigrationTransferId(1)
+            },
+            "the kernel's rebuild stands"
+        );
+        assert_eq!(
+            state.transactions()[1].unsatisfiable_at(),
+            None,
+            "expiry is derived, never recorded"
+        );
+        assert_eq!(store.queries.get(), 1);
+        assert_eq!(
+            store.replaced.get(),
+            0,
+            "nothing was recorded, so nothing is written"
+        );
+    }
+
+    /// The broaden sweep is scoped to transactions the oracle can actually answer about: one
+    /// whose dependencies have yet to mine is never asked (it could only answer "not yet", and
+    /// its death, if any, arrives through the closure), while its unblocked sibling is.
+    #[test]
+    fn advance_broaden_sweep_skips_a_transaction_behind_an_unmined_dependency() {
+        // A transfer funded by a preparation layer that has not been broadcast yet, alongside one
+        // whose funding preparation is already mined.
+        let mut behind_unmined_dep =
+            scheduled_transfer(3, 1, 1440, 90_000, MigrationTxState::Signed);
+        behind_unmined_dep.depends_on = vec![MigrationTransferId(2)];
+        let mut unblocked = scheduled_transfer(4, 2, 1440, 90_000, MigrationTxState::Signed);
+        unblocked.depends_on = vec![MigrationTransferId(0)];
+        let mut state = state_with_crossings(
+            &[5_000_000, 90_000_000, 5_000_000],
+            vec![
+                tx(0, prep(0, 0), mined(10)),
+                // The candidate: proved, due, and dead.
+                scheduled_transfer(1, 0, 1440, 1500, MigrationTxState::Proved),
+                // A second preparation layer, not yet due, so not itself actionable.
+                {
+                    let mut t = tx(2, prep(1, 0), MigrationTxState::Signed);
+                    t.scheduled_height = BlockHeight::from_u32(90_000);
+                    t
+                },
+                behind_unmined_dep,
+                unblocked,
+            ],
+        );
+        let mut store = TestStore::new(
+            1600,
+            [
+                (MigrationTransferId(1), spent(1600)),
+                // Answered only if asked — and the sweep must not ask.
+                (MigrationTransferId(3), spent(1600)),
+                (MigrationTransferId(4), spent(1600)),
+            ],
+        );
+
+        let step = advance_migration(
+            &mut store,
+            &mut state,
+            DuenessTargets::at(BlockHeight::from_u32(1601)),
+            &config(),
+        )
+        .expect("the store never fails");
+
+        assert_eq!(
+            step,
+            AdvanceStep::Waiting,
+            "the live work is all undue, and the marked share stays below the threshold"
+        );
+        assert_eq!(
+            store.queries.get(),
+            3,
+            "the candidate, then the two transactions the sweep can answer about"
+        );
+        assert_eq!(
+            state.transactions()[3].unsatisfiable_at(),
+            None,
+            "a transaction behind an unmined dependency is never asked, so never marked"
+        );
+        assert_eq!(
+            state.transactions()[4].unsatisfiable_at(),
+            Some(BlockHeight::from_u32(1600)),
+            "its unblocked sibling is swept and marked"
+        );
+        assert_eq!(store.replaced.get(), 1);
+    }
+
+    /// A store error propagates, and the call's contract on the way out: nothing is left half
+    /// written. An oracle failure records nothing and writes nothing; a WRITE failure surfaces the
+    /// error with the discovery live only in the in-memory state, which is exactly why the caller
+    /// must re-read after an error rather than trusting what it holds.
+    #[test]
+    fn advance_store_errors_propagate_without_recording() {
+        let migration = || {
+            state_with_crossings(
+                &[100_000_000],
+                vec![
+                    tx(0, prep(0, 0), mined(10)),
+                    scheduled_transfer(1, 0, 1440, 1500, MigrationTxState::Proved),
+                ],
+            )
+        };
+
+        // The oracle fails on the candidate check.
+        let mut state = migration();
+        let mut store = FailingStore::new(Failure::Oracle, 1600, []);
+        assert_eq!(
+            advance_migration(
+                &mut store,
+                &mut state,
+                DuenessTargets::at(BlockHeight::from_u32(1601)),
+                &config()
+            ),
+            Err(StoreFailed)
+        );
+        assert!(
+            state
+                .transactions()
+                .iter()
+                .all(|t| t.unsatisfiable_at().is_none()),
+            "an unanswered check determines nothing"
+        );
+        assert_eq!(
+            store.replaced.get(),
+            0,
+            "and nothing is written on the way out"
+        );
+
+        // The oracle answers — the candidate is dead, and it is the only checkable transaction —
+        // but persisting the discovery fails.
+        let mut state = migration();
+        let mut store = FailingStore::new(
+            Failure::Write,
+            1600,
+            [(MigrationTransferId(1), spent(1600))],
+        );
+        assert_eq!(
+            advance_migration(
+                &mut store,
+                &mut state,
+                DuenessTargets::at(BlockHeight::from_u32(1601)),
+                &config()
+            ),
+            Err(StoreFailed)
+        );
+        assert_eq!(
+            store.queries.get(),
+            1,
+            "the candidate, and nothing to broaden to"
+        );
+        assert_eq!(store.replaced.get(), 1, "the write was attempted");
+        assert!(
+            state.transactions()[1].unsatisfiable_at().is_some(),
+            "the discovery is in the in-memory state"
+        );
+        assert!(
+            store.stored.is_none(),
+            "but not in the store: after an error the caller must re-read, not trust this state"
+        );
+    }
+
+    /// A rejected broadcast whose explanation lies above the wallet's scanned region: the call
+    /// surfaces `Reevaluate` and nothing else, records nothing, and does not fall through to the
+    /// migration's other due work — a rejection means another observer saw chain state this
+    /// wallet has not, which is a fact about the whole view rather than one transaction.
+    #[test]
+    fn advance_holds_at_reevaluate_until_the_oracle_reaches_the_reported_tip() {
+        let mut state = state_with_crossings(
+            &[50_000_000, 50_000_000],
+            vec![
+                tx(0, prep(0, 0), mined(10)),
+                scheduled_transfer(1, 0, 1440, 1500, MigrationTxState::Proved),
+                scheduled_transfer(2, 1, 1440, 1500, MigrationTxState::Proved),
+            ],
+        );
+        // The node that refused the broadcast reported a tip 100 blocks above the wallet's
+        // fully-scanned height.
+        state.report_broadcast_failure(MigrationTransferId(1), BlockHeight::from_u32(1700));
+        let mut store = TestStore::new(1600, []);
+
+        assert_eq!(
+            advance_migration(
+                &mut store,
+                &mut state,
+                DuenessTargets::at(BlockHeight::from_u32(1601)),
+                &config()
+            )
+            .expect("the store never fails"),
+            AdvanceStep::Reevaluate,
+            "the sibling's due broadcast waits behind the unanswered rejection",
+        );
+        assert_eq!(store.queries.get(), 1, "one query, for the reported row");
+        assert_eq!(store.replaced.get(), 0, "nothing was determined to persist");
+        assert_eq!(
+            state.transactions()[1].broadcast_failure_at(),
+            Some(BlockHeight::from_u32(1700)),
+            "the report stands until it can be answered",
+        );
+        assert!(
+            state
+                .transactions()
+                .iter()
+                .all(|t| t.unsatisfiable_at().is_none()),
+            "testimony never marks",
+        );
+        assert_eq!(
+            state.transaction_statuses(DuenessTargets::at(BlockHeight::from_u32(1601)))[1]
+                .blocked_on(),
+            Some(crate::state::Blocker::AwaitingReevaluation),
+        );
+
+        // The wallet syncs past the reported tip and the same answer now decides the question:
+        // nothing obstructs, so the rejection was transient, the report is discharged, and the
+        // broadcast is offered again IN THIS CALL.
+        store.as_of_height = BlockHeight::from_u32(1700);
+        assert_eq!(
+            advance_migration(
+                &mut store,
+                &mut state,
+                DuenessTargets::at(BlockHeight::from_u32(1701)),
+                &config()
+            )
+            .expect("the store never fails"),
+            AdvanceStep::Broadcast {
+                id: MigrationTransferId(1)
+            },
+        );
+        assert_eq!(state.transactions()[1].broadcast_failure_at(), None);
+        assert_eq!(
+            store
+                .stored
+                .as_ref()
+                .expect("the discharge was persisted")
+                .transactions()[1]
+                .broadcast_failure_at(),
+            None,
+            "the discharge is durable before the step is surfaced",
+        );
+    }
+
+    /// The other adjudication: once the wallet has scanned to the reported tip it finds the spend
+    /// behind the rejection, and that becomes an ORDINARY evidence-backed mark — closure,
+    /// threshold, and `Replan` exactly as any other discovery. The report is discharged either
+    /// way; the mark is what stands afterwards.
+    #[test]
+    fn advance_adjudicates_a_reported_broadcast_into_marks_and_a_replan() {
+        let mut prep_layer_1 = tx(1, prep(1, 0), MigrationTxState::Proved);
+        prep_layer_1.depends_on = vec![MigrationTransferId(0)];
+        prep_layer_1.scheduled_height = BlockHeight::from_u32(1500);
+        let mut first = scheduled_transfer(2, 0, 1440, 90_000, MigrationTxState::Signed);
+        first.depends_on = vec![MigrationTransferId(1)];
+        let mut second = scheduled_transfer(3, 1, 1440, 90_000, MigrationTxState::Signed);
+        second.depends_on = vec![MigrationTransferId(1)];
+
+        let mut state = state_with_crossings(
+            &[50_000_000, 50_000_000],
+            vec![tx(0, prep(0, 0), mined(10)), prep_layer_1, first, second],
+        );
+        state.report_broadcast_failure(MigrationTransferId(1), BlockHeight::from_u32(1700));
+
+        // Another wallet on the same seed spent the notes this preparation was built over, and
+        // the wallet has now scanned the block that did it.
+        let mut store = TestStore::new(1700, [(MigrationTransferId(1), spent(1700))]);
+        assert_eq!(
+            advance_migration(
+                &mut store,
+                &mut state,
+                DuenessTargets::at(BlockHeight::from_u32(1701)),
+                &config()
+            )
+            .expect("the store never fails"),
+            AdvanceStep::Replan,
+        );
+
+        let marked = BlockHeight::from_u32(1700);
+        assert_eq!(
+            state.transactions()[1].unsatisfiable(),
+            Some((marked, UnsatisfiableKind::InputsSpent)),
+            "the rejection is adjudicated into an evidence-backed mark at the answer's height",
+        );
+        assert_eq!(
+            state.transactions()[1].broadcast_failure_at(),
+            None,
+            "and the testimony it rested on is discharged",
+        );
+        for stranded in [2, 3] {
+            assert_eq!(
+                state.transactions()[stranded].unsatisfiable(),
+                Some((marked, UnsatisfiableKind::Inherited)),
+                "the durable closure strands the crossings behind the dead preparation",
+            );
+        }
+        assert!(state.replan_required());
+        assert_eq!(
+            store.replaced.get(),
+            1,
+            "the marks and the discharge persist in one write"
+        );
+        let stored = store.stored.as_ref().expect("persisted");
+        assert_eq!(
+            stored.transactions()[1].unsatisfiable(),
+            Some((marked, UnsatisfiableKind::InputsSpent))
+        );
+        assert_eq!(stored.transactions()[1].broadcast_failure_at(), None);
+    }
+
+    /// A TERMINAL migration is never driven further, so a report left standing on it neither
+    /// costs a query nor holds the driver at `Reevaluate`: there is no broadcast to withhold.
+    #[test]
+    fn advance_ignores_a_report_on_a_terminal_migration() {
+        let mut state = state_with_crossings(
+            &[100_000_000],
+            vec![
+                tx(0, prep(0, 0), mined(10)),
+                scheduled_transfer(1, 0, 1440, 1500, MigrationTxState::Proved),
+            ],
+        );
+        state.report_broadcast_failure(MigrationTransferId(1), BlockHeight::from_u32(1700));
+        state.mark_superseded();
+        let mut store = TestStore::new(1600, []);
+
+        assert_eq!(
+            advance_migration(
+                &mut store,
+                &mut state,
+                DuenessTargets::at(BlockHeight::from_u32(1601)),
+                &config()
+            )
+            .expect("the store never fails"),
+            AdvanceStep::Complete,
+        );
+        assert_eq!(store.queries.get(), 0);
+        assert_eq!(store.replaced.get(), 0);
+    }
+
+    /// `Reevaluate` is a RETURN like any other, so the determinations made on the way to it are
+    /// durable before it is surfaced. Here the in-flight sweep marks a broadcast crossing whose
+    /// anchor a settled reorg displaced, while a second crossing's rejection cannot yet be
+    /// answered: the call ends at `Reevaluate`, and the mark is in the STORE, not only in the
+    /// returned state — a consumer that syncs, restarts, and drives again must not have to
+    /// rediscover it.
+    #[test]
+    fn advance_persists_sweep_marks_before_returning_reevaluate() {
+        let mut in_flight = scheduled_transfer(
+            1,
+            0,
+            1440,
+            1500,
+            MigrationTxState::Broadcast {
+                txid: TxId::from_bytes([4; 32]),
+            },
+        );
+        in_flight.depends_on = vec![MigrationTransferId(0)];
+        let mut reported = scheduled_transfer(2, 1, 1440, 1500, MigrationTxState::Proved);
+        reported.depends_on = vec![MigrationTransferId(0)];
+
+        let mut state = state_with_crossings(
+            &[50_000_000, 50_000_000],
+            vec![tx(0, prep(0, 0), mined(10)), in_flight, reported],
+        );
+        state.report_broadcast_failure(MigrationTransferId(2), BlockHeight::from_u32(1700));
+
+        let mut store = TestStore::new(
+            1600,
+            [(
+                MigrationTransferId(1),
+                StepSatisfiability::Unsatisfiable {
+                    cause: UnsatisfiableCause::AnchorInvalidated,
+                    as_of_height: BlockHeight::from_u32(1600),
+                },
+            )],
+        );
+
+        assert_eq!(
+            advance_migration(
+                &mut store,
+                &mut state,
+                DuenessTargets::at(BlockHeight::from_u32(1601)),
+                &config()
+            )
+            .expect("the store never fails"),
+            AdvanceStep::Reevaluate,
+        );
+        let marked = BlockHeight::from_u32(1600);
+        assert_eq!(
+            state.transactions()[1].unsatisfiable(),
+            Some((marked, UnsatisfiableKind::AnchorInvalidated)),
+        );
+        assert_eq!(
+            store.replaced.get(),
+            1,
+            "one write carries every determination this call made"
+        );
+        assert_eq!(
+            store
+                .stored
+                .as_ref()
+                .expect("the sweep's mark was persisted before the step was surfaced")
+                .transactions()[1]
+                .unsatisfiable(),
+            Some((marked, UnsatisfiableKind::AnchorInvalidated)),
+        );
+    }
+
+    /// The same durability, for the adjudication's own half: with two rejections outstanding and
+    /// only one of them answerable, the answerable one is discharged and that discharge is
+    /// PERSISTED even though the call still ends at `Reevaluate`. Losing it would leave the
+    /// crossing withheld from the broadcast queue on a question already settled.
+    #[test]
+    fn advance_persists_a_discharged_report_before_returning_reevaluate() {
+        let mut answerable = scheduled_transfer(1, 0, 1440, 1500, MigrationTxState::Proved);
+        answerable.depends_on = vec![MigrationTransferId(0)];
+        let mut pending = scheduled_transfer(2, 1, 1440, 1500, MigrationTxState::Proved);
+        pending.depends_on = vec![MigrationTransferId(0)];
+
+        let mut state = state_with_crossings(
+            &[50_000_000, 50_000_000],
+            vec![tx(0, prep(0, 0), mined(10)), answerable, pending],
+        );
+        // The first rejection came from a node whose tip the wallet has since scanned past; the
+        // second from one reporting a tip it has not reached.
+        state.report_broadcast_failure(MigrationTransferId(1), BlockHeight::from_u32(1500));
+        state.report_broadcast_failure(MigrationTransferId(2), BlockHeight::from_u32(1700));
+        let mut store = TestStore::new(1600, []);
+
+        assert_eq!(
+            advance_migration(
+                &mut store,
+                &mut state,
+                DuenessTargets::at(BlockHeight::from_u32(1601)),
+                &config()
+            )
+            .expect("the store never fails"),
+            AdvanceStep::Reevaluate,
+            "one unanswerable rejection still holds the whole loop",
+        );
+        assert_eq!(state.transactions()[1].broadcast_failure_at(), None);
+        assert_eq!(
+            state.transactions()[2].broadcast_failure_at(),
+            Some(BlockHeight::from_u32(1700)),
+        );
+        assert_eq!(store.replaced.get(), 1);
+        let stored = store
+            .stored
+            .as_ref()
+            .expect("the discharge was persisted before the step was surfaced");
+        assert_eq!(
+            stored.transactions()[1].broadcast_failure_at(),
+            None,
+            "the settled question does not withhold the crossing again after a restart",
+        );
+        assert_eq!(
+            stored.transactions()[2].broadcast_failure_at(),
+            Some(BlockHeight::from_u32(1700)),
+            "and the open one still stands",
+        );
+    }
+
+    /// The dual-target contract at the DRIVE level: a wallet woken by wall-clock estimate, its
+    /// scan a couple of hundred blocks behind, submits the transfer whose schedule the estimate
+    /// says has arrived — and is refused the one whose expiry the estimate says has passed. The
+    /// refusal records NOTHING (the whole point of the scanned anchoring: only a reorg truncation
+    /// could withdraw a mark, and an estimate names no chain state to truncate), so the store is
+    /// never written to and the transfer is offered again the moment its schedule and its expiry
+    /// are both judged the same way.
+    #[test]
+    fn advance_serves_the_schedule_by_estimate_and_withholds_a_doomed_broadcast() {
+        let mut doomed = scheduled_transfer(1, 0, 1440, 1500, MigrationTxState::Proved);
+        doomed.expiry_height = BlockHeight::from_u32(1600);
+        doomed.depends_on = vec![MigrationTransferId(0)];
+        let mut live = scheduled_transfer(2, 1, 1440, 1500, MigrationTxState::Proved);
+        live.expiry_height = BlockHeight::from_u32(2000);
+        live.depends_on = vec![MigrationTransferId(0)];
+
+        let mut state = state_with_crossings(
+            &[50_000_000, 50_000_000],
+            vec![tx(0, prep(0, 0), mined(10)), doomed, live],
+        );
+        let mut store = TestStore::new(1501, []);
+        // Scanned to 1500 (target 1501), estimated tip 1700 (target 1701): both transfers are due
+        // by the estimate and neither is due by the scan, and the first one's expiry lies in the
+        // doomed window `1501 <= 1600 < 1701`.
+        let targets = DuenessTargets::new(BlockHeight::from_u32(1501), BlockHeight::from_u32(1701));
+
+        assert_eq!(
+            advance_migration(&mut store, &mut state, targets, &config())
+                .expect("the store never fails"),
+            AdvanceStep::Broadcast {
+                id: MigrationTransferId(2)
+            },
+            "the live transfer is served by the estimate; the doomed one is withheld"
+        );
+        assert_eq!(
+            state.transactions()[1].unsatisfiable(),
+            None,
+            "the withhold is protective, never a determination"
+        );
+        assert_eq!(
+            store.replaced.get(),
+            0,
+            "and so there is nothing to persist"
+        );
+
+        // Withheld, not condemned: an estimate that no longer overshoots the expiry — the wallet
+        // synced, and the transfer had not lapsed after all — offers the same broadcast.
+        state.mark_broadcast(MigrationTransferId(2), TxId::from_bytes([2; 32]));
+        assert_eq!(
+            advance_migration(
+                &mut store,
+                &mut state,
+                DuenessTargets::at(BlockHeight::from_u32(1501)),
+                &config(),
+            )
+            .expect("the store never fails"),
+            AdvanceStep::Broadcast {
+                id: MigrationTransferId(1)
+            },
+        );
+    }
+}
